Shubham Changes Abhishek Commit(20-04-2024  22:03)

+++ src/main/java/com/enach/Models/MandateTypeAmountResponse.java
@@ -2,8 +2,10 @@ package com.enach.Models;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MandateTypeAmountResponse extends CommonResponse{
 
-    private double amount;
+    private BigDecimal amount;
 }
